requirements.txt
tensorflow>=2.15.0
httpx
pytest
kaggle
anyio
tensorflow>=2.15.0
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
pandas
matplotlib
seaborn
scikit-learn
streamlit
requests
locust
opencv-python-headless
plotly
